requirements.txt
cassandra-driver==3.6.0
cassandra-driver-dse==1.0.3
Django==1.10.2
django-smart-selects==1.2.2
django-uuslug==1.1.8
python-slugify==1.2.1
six==1.10.0
Unidecode==0.4.19
